RTree query in tagMesh*Round
